Daemon: use regular binding for auditModule
Do not rely anymore on artificial FakeAuditService.Module for testing
but instead just pass an extra sysTestModule only on the tests that
need to mock the underlying audit system.

This allows to have a proper audit engine for all tests and a
behaviour that is a lot more similar to the real Gerrit server running.

Having a FakeAuditService.Module always included was breaking existing
tests of other audit plugins, e.g. audit-sl4j.

Bug: Issue 10428
Change-Id: I58f8a0e5ffc2a70073ce0f8e3f10d8860715cd86

Former-commit-id: 4780124708cd0b1364e9259f421f5381e0808ed3