Rely on GroupAuditService rather than its implementation
GroupAuditService is the interface to the underlying audit
system, while AuditService is the actual implementation.

Instead of directly injecting the implementation class, rely on
the interface and leave the association to AuditService to the
Guice binding.

This is useful because tests may want to replace, at times, the
actual implementation of the audit-trail with a fake in-memory
store.

Change-Id: I31bbb522fb974c2504635ccf1251287d4890af89
Former-commit-id: 6a130c17a26f776c9b4a9204de9fcc834511176f