Upgrade mockito-core to 2.24.0 and byte-buddy to 1.9.7
mockito-core is used by the delete-project and webhooks plugin
tests, and was upgraded to this version for the standalone build
on their stable-2.16 branches.

The standalone build doesn't exist for those plugins on master
since they were promoted to core plugins, therefore we do the
upgrade in core.

byte-buddy is a transient dependency of mockito-core and is also
upgraded to the corresponding version.

Change-Id: I878a6c1cd73c595cfe6b3b23b6fa83fc2780647d

Former-commit-id: ff55ad00ac796c0df90d1c8bce872ade5f3189fe