AbstractDaemonTest: Add method for temporarily installing plugin
A plugin is really just a name plus one or more module classes. We
already have the TestServerPlugin magic which is used by
LightweightPluginDaemonTest, so reuse that. The difference is that the
use case for this new method is for testing the plugin plumbing itself,
not testing a specific real-world plugin.

Change-Id: I3ca4eda806519ad9bdd653e0b5f897e69ed119af

Former-commit-id: 737fb08e06e2cf8bef826a9c564dc026d7f0084d