ChangeQueryProcessor: Don't rescan plugins when sizing list
DynamicMap#plugins() creates a new TreeSet on every invocation. We were
just doing this to presize an ArrayList, which makes it seem like we
care about reducing allocations, but was probably actually increasing
allocations. Fortunately, we can just presize using the multimap we
already have. This also allows us to remove a redundant instance field.

Change-Id: Ie5998add61d4848e184a37f4f237da132802a00f

Former-commit-id: 5b33f03e5ecd5bb6cc41368495ba16f7d565e10e