Add acceptance test for plugin URL bindings
For now, just tests a top-level binding under /plugins/X. Future tests
could check views bound in a RestApiModule under an existing collection,
but I wasn't able to get those working in this change.

These tests are in a separate class from the old
PluginsRestApiBindingsIT because they are really testing a different
type of endpoint: the former is about the remote admin endpoints, which
is different from actual plugin-provided endpoints. Rename
PluginsRestApiBindingsIT accordingly.

Change-Id: I71b82a14ded11cc77baf1313d82bc714f8602f67

Former-commit-id: 2a6dee2f1ffd2ec2584757b5d18fbaf4d8978a86