Add mechanism for plugins to modify rest api calls
A plugin can registers a rest api hook to add plugin params to core rest api
calls. The addParameterFunction only needs to return plugin specific params,
not the entire params object.

This will slow down the initial call to get changes and the change detail as
the xhr now waits for the plugins to finish loading. However once plugins are
loaded, the xhr will fire as quickly as before.

This is currently only supported in the changes query and change detail.

Change-Id: I894628663c6616152030d05e41a0008d1601af3f

Former-commit-id: 92daa3334225f7b22c7de83620afbb63385b9d56