Expand acceptance tests for ChangeMessageModifiers
Enforce the current behavior of short-circuiting and failing the entire
operation when a single ChangeMessageModifier fails. This is not the
convention elsewhere in Gerrit; typically we log and continue if a
single plugin fails. This also means we can't switch to using
PluginSetContext for the implementation. However, we know that the
Chromium git-numberer plugin depends on this specific behavior (their CI
gets very unhappy if their custom commit footers are missing), so we
can't change it at the moment. In the future, we could extend the
ChangeMessageModifier interface so some plugins are capable of bringing
down the whole operation, while the default is to log-and-ignore. That
is beyond the scope of this change, which, again, is just enforcing the
current behavior.

Change-Id: I5fd1a074868f9248ab018924fb732cae2c30b604

Former-commit-id: 33613895571b0e64a458eea498c8cc3e7d05724b