Start caching top-menus and version
We want to reduce QPS to Gerrit where easily possible. The framework for
this is already in place: On every Response, we can set caching.

We don't want caching to impact users by showing them stale data that is
in the critical path of their work. For admins, this is slightly more
acceptable.

To start with, we cache /server/version (30s) and /server/top-menus
(300s). These only change if Gerrit gets upgraded or plugins that
contribute top-menus reload.

Change-Id: I33ef5792bffba79bde5d7e7dda6b4291f2db9cc7

Former-commit-id: 05e7f308101b0a0fbac2dbd1674f3eed8ec15906