Fix casing of change states for /related endpoint
Our REST documentation mentions that the status contained in
RelatedChangeAndCommitInfo has one of the values NEW, MERGED,
and ABANDONED. It explicitly lists them in upper case. In addition,
our frontend depends on the exact casing of those values.

This regression was introduced as a side-effect of I9ed389b01. Add a
test to prevent future regressions.

Bug: Issue 10559
Change-Id: I7486b71899c57e185d90d7357873834539933637

Former-commit-id: 1da1862f0d0c6ce3296287cdcfc3821de6cc66fb