Consume JGit artifacts from Maven Central
The recent JGit releases were published to the Eclipse repository
due to a security embargo. The embargo has since been lifted, and
the releases are available on Maven Central, so switch back to
consuming the artifacts from there.

This update was already done on the stable branches stable-2.9 to
stable-2.12 by change I054cc91f4, but was lost during the merge to
stable-2.13 in change Ib5243b6d2.

Change-Id: I55bbff5b0c738e9f419949d29b68b7cc88a748a7

Former-commit-id: 7fe139af2d9b40d25cf82b5387b03ad15bc582fd