CmdLineParser: Require plugin option names to start with '-'
This is normally a requirement for args4j options, but in the case of
plugin options, the option name is produced by prefixing with
"--myplugin", so an incorrect name in the @Option annotation in a plugin
will not produce an error. Add a manual check for this case, forcing
options to always be named "--myplugin-o" or "--myplugin--opt".

Change-Id: I9e6810ec5470f001b8e772f046e4d3e8be5f7247

Former-commit-id: 3062b663688741b85a7e74e05c6864278a2280b3