Bind ChangeAttributeFactory as a DynamicSet
DynamicMaps allow a view of plugin-provided bindings keyed by export
names. The main use case for an export name is to provide something with
that name that is visible to an end user, such as a new SSH command. In
the case of ChangeAttributeFactory, the export name is entirely unused;
it's just an arbitrary string that the plugin author provides, which has
to be unique to their plugin, but is otherwise meaningless.

By using DynamicSet instead of DynamicMap for these bindings, we
eliminate one line of boilerplate for plugin authors, and avoid
confusing them with the impression that this arbitrary string has
meaning.

The good news is that old plugins that use an explicit, arbitrary export
name will continue to work, because DynamicSet is implemented internally
by using separate binding annotations per instance. If callers use the
DynamicSet interface, they get a unique annotation for free, but it
still works if they explicitly pass the export annotation.

Convert most tests in PluginFieldsIT to demonstrate the new way, but
leave one test that uses the old explicit export to ensure that it still
works.

Change-Id: Iaf7e6ac8a53be115f1bc01e716e0fb3c04033754

Former-commit-id: 9d2894f966a0c6e96e133ba623e841054e435c18