Move plugin field tests into respective packages
These originated in the "api" package but the REST/SSH tests are a
better fit for the "rest"/"ssh" packages. Similarly, move the abstract
test class to a common place where we don't have to @Ignore it.

Arguably, the SSH tests belong in a separate ssh.change subpackage, but
that doesn't currently exist.

Change-Id: I2feae1edc020c527d089c4d8679d32b85c76744a

Former-commit-id: b7f9409cd0790ee724da1c0ab74fe245e5398fc6