Fix the PolyGerrit run-server.sh script
The go soy parser is not able to handle newer syntax changes and was
deemed to not be worth fixing. So this change removes the dependency
on the go soy parser by instead patching the server response when it
comes back.

Also removes that unused "prod" flag, which allowed pointing to a dist
directory that does not exist anymore.

Also renames some variables in the server.go script.

Tested by checking that tests are run and the proxy is usable:
http://localhost:8081/
http://localhost:8081/elements/core/gr-search-bar/gr-search-bar_test.html

Bug: Issue 10579
Change-Id: I1d2a155e93c9eb42ff72bfa56819f44a266f8fc5

Former-commit-id: b60db4ee2e019cd89ce3b68b494558d790195f7d