Fix contrib/mitm-ui by serving test components
With this fix in place the script mitm-serve-app-dev.sh works
as advertised. It proxies the https prod server and even allows
login. It is fairly slow though compared to run-server.sh, so
it cannot fully replace it at this point.

Note that there is also a WIP Change I0ff7e314a for getting us
from mitmproxy 2.0.2 to 4.0.4.

Change-Id: Id79510228d6e701f724cc947146768fabae58457

Former-commit-id: 618c4fe47093fb288d230502d37fc73c642ed5b0