QueryBuilder: Remove support for AND/OR within a single field operator
Reading the grammar definition and passing in various inputs, I
currently don't think it's possible to get an AND/OR node under a
FIELD_NAME node. In the previous commit, there is now a test for exactly
the construction that is mentioned in this comment ("foo:(a b c)"), and
it shows that this construction is not accepted by the parser.

If I'm wrong and this is somehow supported, than this commit can be
reverted along with an actual example of the parse succeeding.

Change-Id: Id39b3862f254fc2d7355a05c5cf02dc483ab1e7f

Former-commit-id: 37fe9cc1f553d96228b39d799a6d6cc45b0ee181