Increase autocomplete limit for repo inheritance in access page to 50
The backend results are alphabetically sorted, so it can happen that the
exact match is cut off. Of course, there is still a chance of this to
happen with a limit of 50, but it is much less likely, and the proper
fix is much more work. And this fix is extremely simple and easy.

Bug: Issue 10696
Change-Id: I551880b45b2e773e63d7435af4ff4f16c1bc7b9c

Former-commit-id: 3e364262e18afda045be8d42ae9d7802ee6d5b97