Fix plugin-provided change query operators and add test
I9a4088d2 was intended to be a no-op refactoring, but unfortunately it
broke the implementation of plugin-provided change query operators,
which were previously untested. Add a test of the behavior as documented
in dev-plugins.txt.

The test failure was caused by a private method that mutates the
opFactories map, which is intended to be immutable for the lifetime of a
ChangeQueryBuilder. The old implementation wasn't broken per se, because
the private method was called only from the constructor, but it was not
trivial to reason about (hence why the subsequent breakage slipped
through review). Instead of reinstating the mutable field, push the
DynamicMap iteration into the QueryBuilder constructor.

In order to prove that the DynamicMap iteration is typesafe, introduce a
new type parameter to QueryBuilder, allowing us to remove the warning
suppression on the constructor.

Change-Id: I06cf253c82bf03bc97dcba08ca7d494011c88384

Former-commit-id: b53581be87466d663e8c1e8058e58a9a97c2dfb4