Elasticsearch: Update elasticsearch-rest-client to 7.0.0
Hence update httpasyncclient to 4.1.4 and httpcore-nio to 4.4.11.

Also, update the ElasticContainer test version for V7_0 to 7.0.0.

This is the outstanding breaking change part from [1] or the 7.0 release
notes: "Indices created in Elasticsearch 5.x or before will need to be
reindexed with Elasticsearch 6.x in order to be readable by
Elasticsearch 7.x.". So this is being added to the Elasticsearch
compatibility home page through a companion change.

[1] https://www.elastic.co/guide/en/elasticsearch/reference/7.0/breaking-changes-7.0.html

Change-Id: I1cd6048069fdcd0ac28b253b26451d0ad14db5da

Former-commit-id: 32470fdca20bad17135b918da71836d3b777cf57