Record the number of updates to a NoteDb ref
Eventually we may want to limit the total length of a NoteDb ref's
history, to avoid having to scan and store an unbounded amount of commit
data into ChangeNotes. To do that, we first need to record it.

In most cases, this number will equal getChangeMessages().size(), but
that's partly an artifact of the current APIs, which always accompany a
change with a ChangeMessage. The underlying storage layer doesn't
require it.

Don't bump the ChangeNotesCache version number: old entries will just
return an update count of 0. My current thinking is that populating this
field is not worth flushing the persistent cache. If we change our
minds, we can trivially bump the version number.

Change-Id: I9fbcdfa0583a525e60b75672032f3179e5da3c79

Former-commit-id: 10edb7cad80261f77629ad26d70abb424e8f2453